:sparkles: avatar children should be string
string length 에 따른 font size 조정
